rate legend working well on normal topics; fixed num of colors; fixed legend special items col
