+++ cspell.config.json
@@ -510,7 +510,6 @@
 		"pytest",
 		"QPS",
 		"quantile",
-		"quantize",
 		"Race_eth_NHMult_Oth",
 		"racialization",
 		"radm",
